+++ app/src/main/java/com/github/handioq/fanshop/ui/catalog/CatalogMvp.java
@@ -10,7 +10,7 @@ public interface CatalogMvp {
 
     interface Model extends Mvp.Model {
 
-        void getProducts(int category, int offset, int count);
+        void getProducts(long category, int offset, int count);
 
         void setCallback(Callback callback);
 
@@ -38,7 +38,7 @@ public interface CatalogMvp {
 
     interface Presenter extends Mvp.Presenter<CatalogMvp.View> {
 
-        void getProducts(int category, int offset, int limit);
+        void getProducts(long category, int offset, int limit);
 
     }
 }
